missing character in index.html breaking app
In line #27 there is missing the character 'e' for the function. The script will not work. The example/documentation shows it correct but the source file is missing that important data which will break the script. Please be patient with me as I am not a developer so I have very minimalistic dev skills. Let me know if this helps or I am totally wrong. Good luck and thanks for listening.

phaenomenon